requirements.txt
python-dotenv
jinja2
html2text
email-validator
tqdm
